refactor(ui): consolidate duplicated time formatters
formatTimeAgo existed as 7 near-identical copies with diverging behavior
(day caps vs hour+minute, Unknown vs unknown, NaN leakage). One shared
implementation in lib/format-time-ago.ts now serves all sites, with a
long/compact style option and a now override for the two call sites that
pass a client-only clock to avoid hydration mismatches. Small display
normalizations: hours include minutes, ages over 24h cap at days,
invalid dates render unknown instead of NaNd ago, future timestamps
render just now instead of negative minutes.

The stargazer formatTime/formatDate pair existed as 8 copies; the 6
identical ones consolidate into lib/stargazer/format.ts with zero
behavior change. Two intentionally divergent copies (SkyEvents adds the
year; the stargazer page uses 12-hour format on ISO strings) stay local
with comments explaining why.

Findings B2, B3 in planning/codebase-review-2026-06-09.md.
