+++ app/earth-sciences/quake-world-map.tsx
@@ -15,6 +15,7 @@
 
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { cn } from '@/lib/utils';
+import { formatTimeAgo } from '@/lib/format-time-ago';
 import type { ClientEarthquake } from './earth-sciences-client';
 import { safeExternalUrl } from '@/lib/safe-url';
 import {
@@ -39,20 +40,6 @@ function dotColor(mag: number): string {
 
 function dotRadius(mag: number): number {
   return Math.max(2.5, Math.min(mag * 1.4, 14));
-}
-
-function formatTimeAgo(iso: string): string {
-  const then = new Date(iso).getTime();
-  if (!Number.isFinite(then)) return 'unknown';
-  const diffMs = Date.now() - then;
-  if (diffMs < 0) return 'just now';
-  const mins = Math.floor(diffMs / 60_000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  return `${days}d ago`;
 }
 
 export interface QuakeWorldMapProps {
